+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes relevant to players in this project will be documented in th
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 
 ## [Unreleased]
+### Fixed
+- Barracks Spy and Defense Calculator will now round home and away numbers consistently (and avoid floating point errors)
 
 ## [1.47.0] - 2025-10-31
 ### Changed