+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
   - Scholar now requires 2 unlocked techs to unlock (from 7,500 research points)
   - Scion now requires 6 successful attacks to unlock (from 500 prestige)
 
+### Fixed
+- Barracks Spy and Defense Calculator will now round home and away numbers consistently (and avoid floating point errors)
+
 ## [1.47.0] - 2025-10-31
 ### Changed
 - Explore platinum cost formula adjusted